Skunkworks app-manager (#341)
This is work in progress that uses a backend module not known to the core platform. I've added a `hasInterface` check so that this page does not appear if the relevant interface is not provided by the FOLIO backend, so the change will be invisible except (for now) on my dev box.

Why, then, am I merging this into `master`? Because the changes it entails -- to the package file, translations, and settings index -- are just the kind that generate merge conflicts as other work happens in other branches. I would much rather get these in now than spend the rest of my life wrangling trivial comma differences, not to mention the more serious conflicts that could arise in the actual code. Once that's done, further development of the App Manager can continue in a long-lived branch.

(Also, I need to get collaborators to run this code, and that's easier if it's in the main branch.)