UID-138 provide tools to manipulate token expiration dates
Tokens are stored in http-only cookies so we cannot manipulate them
directly, but we do learn about the tokens' expiration dates in the
response body from login and from refresh, and keep track of those
values in redux so we can now (without conducting a failing fetch) when
a cookie is likely to expired.

This tool provides hooks to set the redux values for both tokens to
expired dates, effectively tricking the refresh token rotation process
to fire.

Refs UID-138, FOLIO-3627